Instead of sending literal strings over a serial connection, send bytes.
We send 4 byte packets over the serial connection so the client /
receiving end of the connection doesn't have to worry about packet
assembly and delimiters. We can simply read in Big Endian, buffered
4 byte segments and assign action (byte 0), modifiers (bytes 1-2) and
keycap (byte 3).

+++ windows/variables_windows.go
@@ -20,7 +20,7 @@ type MSG struct {
 }
 
 type KeyState struct {
-	KeyCode int
+	KeyCode uint8
 }
 
 type WindowsCalls struct {
